tests/Document-getElementById.html

<!DOCTYPE html>
<title>Document.getElementById with an empty string</title>
<script src="testharness.js"></script>
<body>
<div id="log"></div>
<div id=""></div>
<script>
test(function() {
  assert_equals(document.getElementById(""), null)
  // XXX needs more tests
})
</script>
